# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanban Retail - Captura de Demanda API",
    "description": "API para gestionar la demanda insatisfecha de una tienda (single-tenant).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/demands": {
      "get": {
        "tags": [
          "demands"
        ],
        "summary": "List Demands",
        "operationId": "list_demands_api_demands_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DemandStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "name": "assigned_to_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned To Id"
            }
          },
          {
            "name": "supplier_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Busca en el nombre libre del producto.",
              "title": "Q"
            },
            "description": "Busca en el nombre libre del producto."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Numero de pagina (1-indexado).",
              "default": 1,
              "title": "Page"
            },
            "description": "Numero de pagina (1-indexado)."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Tamano de pagina.",
              "default": 20,
              "title": "Size"
            },
            "description": "Tamano de pagina."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DemandRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "demands"
        ],
        "summary": "Create Demand",
        "operationId": "create_demand_api_demands_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/demands/reorder": {
      "patch": {
        "tags": [
          "demands"
        ],
        "summary": "Reorder",
        "description": "Reordena tarjetas (y mueve de columna) en bloque.\n\nSe declara antes de las rutas con `/{demand_id}` para que \"reorder\" no se\ninterprete como un id numerico.",
        "operationId": "reorder_api_demands_reorder_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Reorder Api Demands Reorder Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demands/{demand_id}": {
      "get": {
        "tags": [
          "demands"
        ],
        "summary": "Get Demand",
        "operationId": "get_demand_api_demands__demand_id__get",
        "parameters": [
          {
            "name": "demand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Demand Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "demands"
        ],
        "summary": "Update Demand",
        "operationId": "update_demand_api_demands__demand_id__patch",
        "parameters": [
          {
            "name": "demand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Demand Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "demands"
        ],
        "summary": "Delete Demand",
        "operationId": "delete_demand_api_demands__demand_id__delete",
        "parameters": [
          {
            "name": "demand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Demand Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demands/{demand_id}/history": {
      "get": {
        "tags": [
          "demands"
        ],
        "summary": "Get History",
        "operationId": "get_history_api_demands__demand_id__history_get",
        "parameters": [
          {
            "name": "demand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Demand Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusHistoryRead"
                  },
                  "title": "Response Get History Api Demands  Demand Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demands/{demand_id}/status": {
      "patch": {
        "tags": [
          "demands"
        ],
        "summary": "Change Status",
        "description": "Cambia el estado validando la maquina de estados y la version optimista.",
        "operationId": "change_status_api_demands__demand_id__status_patch",
        "parameters": [
          {
            "name": "demand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Demand Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandStatusPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demands/{demand_id}/consolidate": {
      "post": {
        "tags": [
          "demands"
        ],
        "summary": "Consolidate",
        "description": "Consolida un interes adicional de cliente sobre una demanda existente.",
        "operationId": "consolidate_api_demands__demand_id__consolidate_post",
        "parameters": [
          {
            "name": "demand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Demand Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsolidateInterest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsolidateInterest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "ConsolidateInterest"
      },
...
      "DemandCreate": {
        "properties": {
          "product_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "product_name_free": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name Free"
          },
          "variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          },
          "reason": {
            "$ref": "#/components/schemas/DemandReason",
            "default": "OUT_OF_STOCK"
          },
          "channel": {
            "$ref": "#/components/schemas/DemandChannel",
            "default": "STORE"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority",
            "default": "MEDIUM"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "customer_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Contact"
          },
          "customer_consent": {
            "type": "boolean",
            "title": "Customer Consent",
            "default": false
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Url"
          }
        },
        "type": "object",
        "title": "DemandCreate"
      },
      "DemandRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "product_name_free": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name Free"
          },
          "variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "customer_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Contact"
          },
          "customer_consent": {
            "type": "boolean",
            "title": "Customer Consent"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Url"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "product_name_free",
          "variant",
          "quantity",
          "reason",
          "channel",
          "status",
          "priority",
          "notes",
          "customer_contact",
          "customer_consent",
          "assigned_to_id",
          "supplier_id",
          "evidence_url",
          "sort_order",
          "version",
          "opportunity_id",
          "created_at",
          "updated_at"
        ],
        "title": "DemandRead"
      },
...
      "DemandStatus": {
        "type": "string",
        "enum": [
          "NUEVA",
          "VALIDANDO",
          "COTIZANDO",
          "POR_PEDIR",
          "EN_CAMINO",
          "DISPONIBLE",
          "CERRADA",
          "DESCARTADA"
        ],
        "title": "DemandStatus",
        "description": "Estados del tablero Kanban de captura de demanda."
      },
      "DemandStatusPatch": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/DemandStatus"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "description": "Version esperada de la tarjeta."
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status",
          "version"
        ],
        "title": "DemandStatusPatch",
        "description": "Cambio de estado con verificacion de version optimista."
      },
      "DemandUpdate": {
        "properties": {
          "product_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "product_name_free": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name Free"
          },
          "variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DemandReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "channel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DemandChannel"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Priority"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "customer_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Contact"
          },
          "customer_consent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Consent"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Url"
          }
        },
        "type": "object",
        "title": "DemandUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_DemandRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DemandRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[DemandRead]"
      },
...
      "ReorderPatch": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReorderItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ReorderPatch"
      },
...
    }
  }
}